fix: removed bundled wizcli
