+++ .projen/tasks.json
@@ -42,14 +42,6 @@
       ],
       "condition": "! git log --oneline -1 | grep -q \"chore(release):\""
     },
-    "bundle-wiz": {
-      "name": "bundle-wiz",
-      "steps": [
-        {
-          "exec": "ts-node projenrc/bundle-wiz.ts"
-        }
-      ]
-    },
     "clobber": {
       "name": "clobber",
       "description": "hard resets to HEAD of origin and cleans the local repo",
@@ -106,9 +98,6 @@
       "steps": [
         {
           "exec": "ts-node --project tsconfig.dev.json .projenrc.ts"
-        },
-        {
-          "spawn": "bundle-wiz"
         }
       ]
     },
