+++ package.json
@@ -7,7 +7,6 @@
   "scripts": {
     "build": "npx projen build",
     "bump": "npx projen bump",
-    "bundle-wiz": "npx projen bundle-wiz",
     "clobber": "npx projen clobber",
     "compat": "npx projen compat",
     "compile": "npx projen compile",
@@ -49,11 +48,11 @@
     "eslint-plugin-import": "^2.29.0",
     "jest": "^29.7.0",
     "jest-junit": "^15",
-    "jsii": "~5.0.0",
+    "jsii": "~5.3.0",
     "jsii-diff": "^1.92.0",
     "jsii-docgen": "^10.1.4",
     "jsii-pacmak": "^1.92.0",
-    "jsii-rosetta": "~5.0.0",
+    "jsii-rosetta": "~5.3.0",
     "mock-fs": "^5.2.0",
     "npm-check-updates": "^16",
     "projen": "^0.77.0",
