+++ src/types.ts
@@ -1,0 +1,8 @@
+
+export enum Severity {
+  MEDIUM = 'MEDIUM',
+  LOW = 'LOW',
+  INFO = 'INFO',
+  HIGH = 'HIGH',
+  CRITICAL = 'CRITICAL',
+}
